pyproject.toml
[project]
name = "email-processor"
version = "0.1.0"
description = "Email processing system with AI classification"
authors = [
    {name = "Gunnar Raetsch"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.121.0,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "imapclient (>=3.0.1,<4.0.0)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "markdownify (>=1.2.0,<2.0.0)",
    "pdfplumber (>=0.11.7,<0.12.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "python-pptx (>=1.0.2,<2.0.0)",
    "striprtf (>=0.0.29,<0.0.30)",
    "icalendar (>=6.3.2,<7.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pgvector (>=0.3.6,<0.4.0)",
    "numpy (>=1.26.0,<2.0.0)",
    "langchain (>=1.0.5,<2.0.0)",
    "langchain-openai (>=1.0.2,<2.0.0)",
    "langchain-anthropic (>=1.0.2,<2.0.0)",
    "openai (>=2.7.1,<3.0.0)",
    "anthropic (>=0.72.0,<0.73.0)",
    "cryptography (>=41.0.0,<43.0.0)",
    "jsonschema (>=4.23.0,<5.0.0)",
    "pandas (>=2.0.0,<3.0.0)",
    "google-api-python-client (>=2.0.0,<3.0.0)",
    "google-auth (>=2.0.0,<3.0.0)",
    "gspread (>=6.0.0,<7.0.0)",
    "PyJWT (>=2.8.0,<3.0.0)",
    "mcp (>=1.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "mlx-lm (>=0.28.4,<0.29.0)",
    "playwright (>=1.57.0,<2.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "reportlab (>=4.4.6,<5.0.0)",
    "msal (>=1.31.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "faker (>=37.12.0,<38.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "ruff (>=0.14.4,<0.15.0)",
    "mypy (>=1.18.2,<2.0.0)"
]
test = [
    "pytest-mock (>=3.15.1,<4.0.0)"
]
